+++ packages/graphql-codegen-cli/tests/cli-flags.spec.ts
@@ -4,7 +4,8 @@ import { TempDir } from './utils.js';
 const mockConfig = (str: string, file = './codegen.yml') => temp.createFile(file, str);
 const createArgv = (str = ''): string[] => {
   const result = ['node', 'fake.js'];
-  const regexp = /([^\s'"]+(['"])([^\2]*?)\2)|[^\s'"]+|(['"])([^\4]*?)\4/gi;
+  // eslint-disable-next-line no-control-regex
+  const regexp = /([^\s'"]+(['"])([^\x02]*?)\x02)|[^\s'"]+|(['"])([^\x04]*?)\x04/gi;
 
   let match;
   do {